dwarf_parser: handle signed enums sliiiightly better.
Inspired by a bug I had to fix in Humility recently, I noticed that our
handling of core::cmp::Ordering claimed that the value of Less was
0xFFFF_FFFF_FFFF_FFFF. For a 1-byte enum. So that's obviously wrong.

Really, for best user experience / clarity, enum discriminators should
be stored with type information. Nothing in DWARF actually requires all
the discriminators in an enum to be of the same type, hilariously.

But for now, I'm at least masking the unsigned representation to match
the byte size of the enum.